+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 - changed : use specific icon for preview
+- changed : use two rows for datatables and make column fixed width
 - fixed : remove stat tab menu border on dark mode
 - fixed : remove mouse cursor when datatables not clickable
 - fixed : tab nb item update from child component with store